Delete .run directory
This was my attempt at running all the systems with 1 button click. Apparently, IntelliJi is worse than eclipse at this because you can not control the order in which each main method runs. What we can do to do this in IntelliJi is to download a plugin such as https://github.com/rkhmelyuk/multirun which will run each main method in a specific order. This is required as the Host must start before the Client or the Client will not be able to receive anything and end up deadlocked waiting for a packet that will never arrive.